Automobile Characteristics

Cars dataset with features including make, model, year, engine, and other properties of the car used to predict its price.

cars

Format

A data frame with 11914 rows and 16 variables. The variables are as follows:

make

car brand

model

model given by its brand

year

year of manufacture

engine_fuel_type

type of fuel required by its manufacturer

engine_hp

engine horse power

engine_cylinders

number of cylinders

transmission_type

automatic vs. manual

driven_wheels

AWD, FWD, AWD

number_of_doors

Number of Doors

market_category

Luxury, Performance, Hatchback, etc.

vehicle_size

Compact, Midsize, Large

vehicle_style

Type of Vehicle: Sedan, SUV, Coupe, etc.

highway_mpg

highway miles per gallon

city_mpg

city miles per gallon

popularity

Popularity index

msrp

manufacturer's suggested retail price

Source

Taken from Kaggle https://www.kaggle.com/CooperUnion/cardataset.

Details

This package contains a detailed car dataset.

Examples

summary(cars)

#>          make                     model            year     
#>  Chevrolet :1123   Silverado 1500    :  156   Min.   :1990  
#>  Ford      : 881   Tundra            :  140   1st Qu.:2007  
#>  Volkswagen: 809   F-150             :  126   Median :2015  
#>  Toyota    : 746   Sierra 1500       :   90   Mean   :2010  
#>  Dodge     : 626   Beetle Convertible:   89   3rd Qu.:2016  
#>  Nissan    : 558   Tacoma            :   80   Max.   :2017  
#>  (Other)   :7171   (Other)           :11233                 
#>                        engine_fuel_type   engine_hp      engine_cylinders
#>  regular unleaded              :7172    Min.   :  55.0   Min.   : 0.000  
#>  premium unleaded (required)   :2009    1st Qu.: 170.0   1st Qu.: 4.000  
#>  premium unleaded (recommended):1523    Median : 227.0   Median : 6.000  
#>  flex-fuel (unleaded/E85)      : 899    Mean   : 249.4   Mean   : 5.629  
#>  diesel                        : 154    3rd Qu.: 300.0   3rd Qu.: 6.000  
#>  electric                      :  66    Max.   :1001.0   Max.   :16.000  
#>  (Other)                       :  91    NA's   :69       NA's   :30      
#>         transmission_type           driven_wheels  number_of_doors
#>  AUTOMATED_MANUAL: 626    all wheel drive  :2353   Min.   :2.000  
#>  AUTOMATIC       :8266    four wheel drive :1403   1st Qu.:2.000  
#>  DIRECT_DRIVE    :  68    front wheel drive:4787   Median :4.000  
#>  MANUAL          :2935    rear wheel drive :3371   Mean   :3.436  
#>  UNKNOWN         :  19                             3rd Qu.:4.000  
#>                                                    Max.   :4.000  
#>                                                    NA's   :6      
#>            market_category  vehicle_size          vehicle_style 
#>  N/A               :3742   Compact:4764   Sedan          :3048  
#>  Crossover         :1110   Large  :2777   4dr SUV        :2488  
#>  Flex Fuel         : 872   Midsize:4373   Coupe          :1211  
#>  Luxury            : 855                  Convertible    : 793  
#>  Luxury,Performance: 673                  4dr Hatchback  : 702  
#>  Hatchback         : 641                  Crew Cab Pickup: 681  
#>  (Other)           :4021                  (Other)        :2991  
#>   highway_mpg        city_mpg        popularity        msrp        
#>  Min.   : 12.00   Min.   :  7.00   Min.   :   2   Min.   :   2000  
#>  1st Qu.: 22.00   1st Qu.: 16.00   1st Qu.: 549   1st Qu.:  21000  
#>  Median : 26.00   Median : 18.00   Median :1385   Median :  29995  
#>  Mean   : 26.64   Mean   : 19.73   Mean   :1555   Mean   :  40595  
#>  3rd Qu.: 30.00   3rd Qu.: 22.00   3rd Qu.:2009   3rd Qu.:  42231  
#>  Max.   :354.00   Max.   :137.00   Max.   :5657   Max.   :2065902  
#>