Contraceptive method choice

This dataset is a subset of the 1987 National Indonesia Contraceptive Prevalence Survey.

contraception

Format

A data frame with 1473 rows and 10 variables:

wife_age

integer.

wife_education

integer. 1=low, 2, 3, 4=high.

number_of_children_ever_born

integer.

wife_religion

factor. Non-Islam, Islam.

standard_of_living

integer. 1=low, 2, 3, 4=high.

media_exposure

factor. Good, Not Good.

contraceptive

factor. No-use, Long-term, Short-term.

wife_now_working

factor. Yes, No.

husband_education

integer. 1=low, 2, 3, 4=high.

husband_occupation

integer. 1,2,3,4.

Source

Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science

Details

The samples are married women who were either not pregnant or do not know if they were at the time of interview. The problem is to predict the current contraceptive method choice (no use, long-term methods, or short-term methods) of a woman based on her demographic and socio-economic characteristics.

Examples

summary(contraception)

#>     wife_age     wife_education  number_of_children_ever_born   wife_religion 
#>  Min.   :16.00   Min.   :1.000   Min.   : 0.000               Non-Islam: 220  
#>  1st Qu.:26.00   1st Qu.:2.000   1st Qu.: 1.000               Islam    :1253  
#>  Median :32.00   Median :3.000   Median : 3.000                               
#>  Mean   :32.54   Mean   :2.959   Mean   : 3.261                               
#>  3rd Qu.:39.00   3rd Qu.:4.000   3rd Qu.: 4.000                               
#>  Max.   :49.00   Max.   :4.000   Max.   :16.000                               
#>  standard_of_living  media_exposure    contraceptive wife_now_working
#>  Min.   :1.000      Good    :1364   No-use    :629   Yes: 369        
#>  1st Qu.:3.000      Not good: 109   Long-term :333   No :1104        
#>  Median :3.000                      Short-term:511                   
#>  Mean   :3.134                                                       
#>  3rd Qu.:4.000                                                       
#>  Max.   :4.000                                                       
#>  husband_education husband_occupation
#>  Min.   :1.00      Min.   :1.000     
#>  1st Qu.:3.00      1st Qu.:1.000     
#>  Median :4.00      Median :2.000     
#>  Mean   :3.43      Mean   :2.138     
#>  3rd Qu.:4.00      3rd Qu.:3.000     
#>  Max.   :4.00      Max.   :4.000