CIA World Factbook

Demographic and economic data on 197 countries in 2020.

countries

Format

A data frame with 197 rows and 79 variables:

Country.Name

(character)

Country.Code

2 letter ISO Code.

Government.Type

(character) Basic form of government.

Date.of.Founding.Independence

(Date) yyy-mm-dd

Latitude.of.Capital

(character) e.g. "69 11 E"

Longitude.of.Capital

(character) e.g. "34 31 N"

Telephones.fixed.lines

Total number of fixed telephone lines in use.

Telephones.mobile.cellular

Total number of mobile cellular telephone subscribers.

Internet.users

Total number of individuals within a country who can access the Internet at home.

Broadband.fixed.subscriptions

Total number of fixed-broadband subscriptions.

GDP.purchasing.power.parity

This entry gives the gross domestic product (GDP) or value of all final goods and services produced within a nation in a given year. A nation's GDP at purchasing power parity (PPP) exchange rates is the sum value of all goods and services produced in the country valued at prices prevailing in the United States in the year noted. This is the measure most economists prefer when looking at per-capita welfare and when comparing living conditions or use of resources across countries.

GDP.real.growth.rate

GDP growth on an annual basis adjusted for inflation and expressed as a percent.

GDP.per.capita.PPP

GDP on a purchasing power parity basis divided by population as of 1 July for the same year.

Gross.national.saving

Gross national saving is derived by deducting final consumption expenditure (household plus government) from Gross national disposable income.Expressed as percent of GDP. A negative number indicates that the economy as a whole is spending more income than it produces.

Industrial.production.growth.rate

Annual percentage increase in industrial production.

Labor.force

Total labor force.

Unemployment.rate

Percent of the labor force that is without jobs.

Distribution.of.family.income.Gini.index

Taxes.and.other.revenues

Total taxes and other revenues received by the national government as a percent of GDP.

Budget.surplus.or.deficit

Difference between national government revenues and expenditures, expressed as a percent of GDP.

Public.debt

Total of all government borrowings less repayments that are denominated in a country's home currency.

Inflation.rate.consumer.prices

Annual percent change in consumer prices compared with the previous year's consumer prices.

Central.bank.discount.rate

Annualized interest rate a country's central bank charges commercial, depository banks for loans to meet temporary shortages of funds.

Commercial.bank.prime.lending.rate

Simple average of annualized interest rates commercial banks charge on new loans, denominated in the national currency, to their most credit-worthy customers.

Stock.of.narrow.money

Compares the total quantity of currency in circulation (notes and coins) plus demand deposits denominated in the national currency, held by nonbank financial institutions, state and local governments, nonfinancial public enterprises, and the private sector of the economy. The national currency units have been converted to US dollars.

Stock.of.broad.money

All of Narrow money, plus the total quantity of time and savings deposits, credit union deposits, institutional money market funds, short-term repurchase agreements between the central bank and commercial deposit banks, and other large liquid assets held by nonbank financial institutions, state and local governments, nonfinancial public enterprises, and the private sector of the economy. National currency units have been converted to US dollars.

Stock.of.domestic.credit

Total quantity of credit, denominated in the domestic currency, provided by financial institutions to the central bank, state and local governments, public non-financial corporations, and the private sector. The national currency units have been converted to US dollars.

Market.value.of.publicly.traded.shares

The latest price per share multiplied by the total number of outstanding shares, cumulated over all companies listed on the particular exchange.

Current.account.balance

Compares a country's net trade in goods and services, plus net earnings, and net transfer payments to and from the rest of the world during the period specified.

Exports

Total US dollar amount of merchandise exports on an f.o.b. (free on board) basis.

Imports

Total US dollar amount of merchandise imports on a c.i.f. (cost, insurance, and freight) or f.o.b. (free on board) basis.

Reserves.of.foreign.exchange.and.gold

Dollar value for the stock of all financial assets that are available to the central monetary authority for use in meeting a country's balance of payments.

Debt.external

Total public and private debt owed to nonresidents repayable in internationally accepted currencies, goods, or services.

Stock.of.direct.foreign.investment.at.home

Cumulative US dollar value of all investments in the home country made directly by residents - primarily companies - of other countries.

Stock.of.direct.foreign.investment.abroad

Cumulative US dollar value of all investments in foreign countries made directly by residents - primarily companies - of the home country.

Electricity.production

Annual electricity generated expressed in kilowatt-hours.

Electricity.consumption

Total electricity generated annually plus imports and minus exports, expressed in kilowatt-hours.

Electricity.exports

Total exported electricity in kilowatt-hours.

Electricity.imports

Total imported electricity in kilowatt-hours.

Electricity.installed.generating.capacity

Electricity.from.fossil.fuels

Capacity of plants that generate electricity by burning fossil fuels (such as coal, petroleum products, and natural gas), expressed as a share of the country's total generating capacity

Electricity.from.nuclear.fuels

Capacity of plants that generate electricity through radioactive decay of nuclear fuel, expressed as a share of the country's total generating capacity

Electricity.from.hydroelectric.plants

Capacity of plants that generate electricity by water-driven turbines, expressed as a share of the country's total generating capacity

Electricity.from.other.renewable.sources

Capacity of plants that generate electricity by using renewable energy sources other than hydroelectric (including, for example, wind, waves, solar, and geothermal), expressed as a share of the country's total generating capacity.

Crude.oil.production

Crude oil produced, in barrels per day.

Crude.oil.exports

Total amount of crude oil exported, in barrels per day.

Crude.oil.imports

Total amount of crude oil imported, in barrels per day.

Crude.oil.proved.reserves

Barrels of petroleum which can be estimated to be commercially recoverable from a given date forward, from known reservoirs and under current economic conditions.

Refined.petroleum.products.production

Total output of refined petroleum products, in barrels per day.

Refined.petroleum.products.consumption

Total consumption of refined petroleum products, in barrels per day.

Refined.petroleum.products.exports

Total exports of refined petroleum products, in barrels per day.

Refined.petroleum.products.imports

Total imports of refined petroleum products, in barrels per day.

Natural.gas.production

Total natural gas produced in cubic meters.

Natural.gas.consumption

Total natural gas consumed in cubic meters.

Natural.gas.exports

Total natural gas exported in cubic meters.

Natural.gas.imports

Total natural gas imported in cubic meters.

Natural.gas.proved.reserves

The quantities of natural gas in cubic meters, which can be estimated to be commercially recoverable from a given date forward, from known reservoirs and under current economic conditions.

Carbon.dioxide.emissions.from.consumption.of.energy

Total amount of carbon dioxide, measured in metric tons, released by burning fossil fuels in the process of producing and consuming energy.

Area

Sum of all land and water areas delimited by international boundaries and/or coastlines.

Military.expenditures

Spending on defense programs for the most recent year available as a percent of gross domestic product (GDP).

Population

Median.age

Population.growth.rate

Average annual percent change in the population, resulting from a surplus (or deficit) of births over deaths and the balance of migrants entering and leaving a country.

Birth.rate

Average annual number of births during a year per 1,000 persons.

Death.rate

Average annual number of deaths during a year per 1,000 population.

Net.migration.rate

Difference between the number of persons entering and leaving a country during the year per 1,000 persons

Maternal.mortality.rate

Annual number of female deaths per 100,000 live births from any cause related to or aggravated by pregnancy or its management.

Infant.mortality.rate

Number of deaths of infants under one year old in a given year per 1,000 live births. This rate is often used as an indicator of the level of health in a country.

Life.expectancy.at.birth

Average number of years to be lived by a group of people born in the same year, if mortality at each age remains constant in the future. Life expectancy at birth is also a measure of overall quality of life in a country.

Total.fertility.rate

Average number of children that would be born per woman if all women lived to the end of their childbearing years and bore children according to a given fertility rate at each age.

Obesity.adult.prevalence.rate

Percent of a country's population considered to be obese (BMI >=30).

Children.under.the.age.of.5.years.underweight

Percent of children under five considered to be underweight. This statistic is an indicator of the nutritional status of a community.

Education.expenditures

Unemployment.youth.ages.15.24

The public expenditure on education as a percent of GDP.

Airports

Total number of airports or airfields recognizable from the air.

Railways

Total route length of the railway network.

Roadways

Total length of the road network and includes the length of the paved and unpaved portions.

Merchant.marine

Total number of privately or publicly owned commercial ships; military ships are not included.

Capital

(character)Name of the seat of government.

Source

Gopal Remesh Dahale, Kaggle Country_data https://www.kaggle.com/gopalrameshdahale/country-data?select=Country_data.csv

Details

Dataset originally from the CIA World Factbook . Variable descriptions are quoted from this website. More detail can be found at these link.

Examples

summary(countries)

#>  Country.Name       Country.Code       Government.Type   
#>  Length:197         Length:197         Length:197        
#>  Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character  
#>                                                          
#>                                                          
#>                                                          
#>                                                          
#>  Date.of.Founding.Independence Latitude.of.Capital Longitude.of.Capital
#>  Length:197                    Length:197          Length:197          
#>  Class :character              Class :character    Class :character    
#>  Mode  :character              Mode  :character    Mode  :character    
#>                                                                        
#>                                                                        
#>                                                                        
#>                                                                        
#>  Telephones.fixed.lines Telephones.mobile.cellular Internet.users     
#>  Min.   :        0      Min.   :7.600e+03          Min.   :     5042  
#>  1st Qu.:    57057      1st Qu.:1.936e+06          1st Qu.:   542694  
#>  Median :   413100      Median :7.976e+06          Median :  2700726  
#>  Mean   :  5036563      Mean   :4.015e+07          Mean   : 17699419  
#>  3rd Qu.:  2794418      3rd Qu.:2.478e+07          3rd Qu.:  9735055  
#>  Max.   :193762000      Max.   :1.474e+09          Max.   :730723960  
#>  NA's   :4              NA's   :3                  NA's   :7          
#>  Broadband.fixed.subscriptions GDP.purchasing.power.parity GDP.real.growth.rate
#>  Min.   :       76             Min.   :4.200e+07           Min.   :-36.500     
#>  1st Qu.:    24478             1st Qu.:1.943e+10           1st Qu.:  1.800     
#>  Median :   286698             Median :6.893e+10           Median :  3.050     
#>  Mean   :  5409959             Mean   :6.607e+11           Mean   :  3.123     
#>  3rd Qu.:  2375346             3rd Qu.:3.464e+11           3rd Qu.:  4.800     
#>  Max.   :378540000             Max.   :2.536e+13           Max.   : 64.000     
#>  NA's   :13                    NA's   :2                   NA's   :3           
#>  GDP.per.capita.PPP Gross.national.saving Industrial.production.growth.rate
#>  Min.   :   700     Min.   :-21.90        Min.   :-26.100                  
#>  1st Qu.:  4000     1st Qu.: 13.97        1st Qu.:  1.125                  
#>  Median : 12750     Median : 20.55        Median :  3.150                  
#>  Mean   : 21083     Mean   : 21.00        Mean   :  3.399                  
#>  3rd Qu.: 29525     3rd Qu.: 26.80        3rd Qu.:  5.500                  
#>  Max.   :139100     Max.   : 50.20        Max.   : 60.300                  
#>  NA's   :3          NA's   :19            NA's   :11                       
#>   Labor.force        Unemployment.rate Distribution.of.family.income.Gini.index
#>  Min.   :     3615   Min.   : 0.30     Min.   :23.20                           
#>  1st Qu.:   777100   1st Qu.: 4.40     1st Qu.:32.17                           
#>  Median :  3657000   Median : 7.05     Median :38.15                           
#>  Mean   : 17707653   Mean   :10.85     Mean   :39.01                           
#>  3rd Qu.: 11920000   3rd Qu.:12.25     3rd Qu.:45.23                           
#>  Max.   :806700000   Max.   :77.00     Max.   :63.20                           
#>  NA's   :4           NA's   :13        NA's   :47                              
#>  Taxes.and.other.revenues Budget.surplus.or.deficit  Public.debt    
#>  Min.   :  2.10           Min.   :-75.700           Min.   :  2.80  
#>  1st Qu.: 17.68           1st Qu.: -5.425           1st Qu.: 36.15  
#>  Median : 26.40           Median : -2.700           Median : 50.70  
#>  Mean   : 28.79           Mean   : -3.846           Mean   : 56.45  
#>  3rd Qu.: 37.10           3rd Qu.: -0.725           3rd Qu.: 70.55  
#>  Max.   :106.70           Max.   : 25.600           Max.   :237.60  
#>  NA's   :3                NA's   :3                 NA's   :6       
#>  Inflation.rate.consumer.prices Central.bank.discount.rate
#>  Min.   :  -0.90                Min.   :-0.500            
#>  1st Qu.:   1.10                1st Qu.: 1.750            
#>  Median :   2.40                Median : 5.000            
#>  Mean   :  11.22                Mean   : 5.815            
#>  3rd Qu.:   5.40                3rd Qu.: 7.530            
#>  Max.   :1087.50                Max.   :29.500            
#>  NA's   :4                      NA's   :40                
#>  Commercial.bank.prime.lending.rate Stock.of.narrow.money Stock.of.broad.money
#>  Min.   : 1.290                     Min.   :4.407e+07     Min.   :7.538e+07   
#>  1st Qu.: 5.258                     1st Qu.:1.540e+09     1st Qu.:1.784e+09   
#>  Median : 8.755                     Median :6.564e+09     Median :6.756e+09   
#>  Mean   :10.641                     Mean   :1.905e+11     Mean   :1.896e+11   
#>  3rd Qu.:14.127                     3rd Qu.:6.256e+10     3rd Qu.:6.250e+10   
#>  Max.   :60.000                     Max.   :8.351e+12     Max.   :8.351e+12   
#>  NA's   :15                         NA's   :15            NA's   :14          
#>  Stock.of.domestic.credit Market.value.of.publicly.traded.shares
#>  Min.   :-2.406e+08       Min.   :0.000e+00                     
#>  1st Qu.: 2.910e+09       1st Qu.:3.102e+09                     
#>  Median : 1.727e+10       Median :2.071e+10                     
#>  Mean   : 6.218e+11       Mean   :5.447e+11                     
#>  3rd Qu.: 1.737e+11       3rd Qu.:2.439e+11                     
#>  Max.   : 2.734e+13       Max.   :2.507e+13                     
#>  NA's   :16               NA's   :80                            
#>  Current.account.balance    Exports             Imports         
#>  Min.   :-4.491e+11      Min.   :0.000e+00   Min.   :6.490e+07  
#>  1st Qu.:-1.600e+09      1st Qu.:1.070e+09   1st Qu.:2.587e+09  
#>  Median :-3.280e+08      Median :6.788e+09   Median :9.774e+09  
#>  Mean   : 2.065e+09      Mean   :8.849e+10   Mean   :8.770e+10  
#>  3rd Qu.: 8.790e+08      3rd Qu.:4.760e+10   3rd Qu.:5.147e+10  
#>  Max.   : 2.910e+11      Max.   :2.490e+12   Max.   :2.361e+12  
#>  NA's   :8               NA's   :3           NA's   :3          
#>  Reserves.of.foreign.exchange.and.gold Debt.external      
#>  Min.   :0.000e+00                     Min.   :0.000e+00  
#>  1st Qu.:6.577e+08                     1st Qu.:2.772e+09  
#>  Median :4.708e+09                     Median :1.748e+10  
#>  Mean   :6.556e+10                     Mean   :3.981e+11  
#>  3rd Qu.:3.342e+10                     3rd Qu.:9.436e+10  
#>  Max.   :3.236e+12                     Max.   :1.791e+13  
#>  NA's   :12                            NA's   :7          
#>  Stock.of.direct.foreign.investment.at.home
#>  Min.   :1.580e+07                         
#>  1st Qu.:6.120e+09                         
#>  Median :3.383e+10                         
#>  Mean   :2.523e+11                         
#>  3rd Qu.:1.796e+11                         
#>  Max.   :5.499e+12                         
#>  NA's   :64                                
#>  Stock.of.direct.foreign.investment.abroad Electricity.production
#>  Min.   :0.000e+00                         Min.   :0.000e+00     
#>  1st Qu.:9.392e+08                         1st Qu.:1.005e+09     
#>  Median :1.630e+10                         Median :1.097e+10     
#>  Mean   :3.084e+11                         Mean   :1.230e+11     
#>  3rd Qu.:1.418e+11                         3rd Qu.:5.938e+10     
#>  Max.   :6.579e+12                         Max.   :5.883e+12     
#>  NA's   :82                                NA's   :5             
#>  Electricity.consumption Electricity.exports Electricity.imports
#>  Min.   :0.000e+00       Min.   :0.000e+00   Min.   :0.000e+00  
#>  1st Qu.:1.202e+09       1st Qu.:0.000e+00   1st Qu.:0.000e+00  
#>  Median :9.363e+09       Median :0.000e+00   Median :8.300e+07  
#>  Mean   :1.132e+11       Mean   :3.479e+09   Mean   :3.600e+09  
#>  3rd Qu.:4.936e+10       3rd Qu.:1.428e+09   3rd Qu.:2.546e+09  
#>  Max.   :5.564e+12       Max.   :7.886e+10   Max.   :7.272e+10  
#>  NA's   :6               NA's   :5           NA's   :5          
#>  Electricity.installed.generating.capacity Electricity.from.fossil.fuels
#>  Min.   :6.000e+02                         Min.   :  0.00               
#>  1st Qu.:5.095e+05                         1st Qu.: 41.00               
#>  Median :2.784e+06                         Median : 65.00               
#>  Mean   :3.383e+07                         Mean   : 61.96               
#>  3rd Qu.:1.658e+07                         3rd Qu.: 86.50               
#>  Max.   :1.653e+09                         Max.   :100.00               
#>  NA's   :6                                 NA's   :6                    
#>  Electricity.from.nuclear.fuels Electricity.from.hydroelectric.plants
#>  Min.   : 0.000                 Min.   :  0.00                       
#>  1st Qu.: 0.000                 1st Qu.:  1.00                       
#>  Median : 0.000                 Median : 16.00                       
#>  Mean   : 1.921                 Mean   : 25.69                       
#>  3rd Qu.: 0.000                 3rd Qu.: 42.00                       
#>  Max.   :50.000                 Max.   :100.00                       
#>  NA's   :6                      NA's   :6                            
#>  Electricity.from.other.renewable.sources Crude.oil.production
#>  Min.   :  0.00                           Min.   :       0    
#>  1st Qu.:  1.00                           1st Qu.:       0    
#>  Median :  5.50                           Median :     160    
#>  Mean   : 10.57                           Mean   :  434078    
#>  3rd Qu.: 16.00                           3rd Qu.:   60500    
#>  Max.   :100.00                           Max.   :10962000    
#>  NA's   :7                                NA's   :6           
#>  Crude.oil.exports Crude.oil.imports Crude.oil.proved.reserves
#>  Min.   :      0   Min.   :      0   Min.   :0.000e+00        
#>  1st Qu.:      0   1st Qu.:      0   1st Qu.:0.000e+00        
#>  Median :      0   Median :      0   Median :1.690e+06        
#>  Mean   : 222494   Mean   : 233577   Mean   :8.618e+09        
#>  3rd Qu.:  23220   3rd Qu.:  65485   3rd Qu.:4.275e+08        
#>  Max.   :7341000   Max.   :7969000   Max.   :3.023e+11        
#>  NA's   :6         NA's   :6         NA's   :9                
#>  Refined.petroleum.products.production Refined.petroleum.products.consumption
#>  Min.   :       0                      Min.   :     400                      
#>  1st Qu.:       0                      1st Qu.:   16750                      
#>  Median :   13120                      Median :   53000                      
#>  Mean   :  469399                      Mean   :  511045                      
#>  3rd Qu.:  222900                      3rd Qu.:  261250                      
#>  Max.   :20300000                      Max.   :19960000                      
#>  NA's   :6                             NA's   :9                             
#>  Refined.petroleum.products.exports Refined.petroleum.products.imports
#>  Min.   :      0                    Min.   :      0                   
#>  1st Qu.:      0                    1st Qu.:   7360                   
#>  Median :   2431                    Median :  31035                   
#>  Mean   : 165565                    Mean   : 150752                   
#>  3rd Qu.:  61660                    3rd Qu.: 114050                   
#>  Max.   :5218000                    Max.   :2335000                   
#>  NA's   :7                          NA's   :9                         
#>  Natural.gas.production Natural.gas.consumption Natural.gas.exports
#>  Min.   :0.000e+00      Min.   :0.000e+00       Min.   :0.000e+00  
#>  1st Qu.:0.000e+00      1st Qu.:0.000e+00       1st Qu.:0.000e+00  
#>  Median :0.000e+00      Median :4.814e+08       Median :0.000e+00  
#>  Mean   :1.938e+10      Mean   :1.963e+10       Mean   :6.139e+09  
#>  3rd Qu.:5.762e+09      3rd Qu.:1.006e+10       3rd Qu.:1.200e+08  
#>  Max.   :7.728e+11      Max.   :7.676e+11       Max.   :2.102e+11  
#>  NA's   :6              NA's   :7               NA's   :7          
#>  Natural.gas.imports Natural.gas.proved.reserves
#>  Min.   :0.000e+00   Min.   :0.000e+00          
#>  1st Qu.:0.000e+00   1st Qu.:0.000e+00          
#>  Median :0.000e+00   Median :2.960e+09          
#>  Mean   :6.237e+09   Mean   :1.019e+12          
#>  3rd Qu.:2.244e+09   3rd Qu.:1.416e+11          
#>  Max.   :1.195e+11   Max.   :4.780e+13          
#>  NA's   :7           NA's   :12                 
#>  Carbon.dioxide.emissions.from.consumption.of.energy      Area         
#>  Min.   :1.050e+02                                   Min.   :       2  
#>  1st Qu.:3.062e+06                                   1st Qu.:   26026  
#>  Median :1.218e+07                                   Median :  130370  
#>  Mean   :1.972e+08                                   Mean   :  698324  
#>  3rd Qu.:7.207e+07                                   3rd Qu.:  554168  
#>  Max.   :1.167e+10                                   Max.   :17098242  
#>  NA's   :8                                           NA's   :2         
#>  Military.expenditures   Population          Median.age   
#>  Min.   :0.160         Min.   :9.692e+03   Min.   :15.50  
#>  1st Qu.:0.980         1st Qu.:1.916e+06   1st Qu.:22.80  
#>  Median :1.440         Median :8.293e+06   Median :29.20  
#>  Mean   :1.857         Mean   :3.833e+07   Mean   :30.08  
#>  3rd Qu.:2.290         3rd Qu.:2.767e+07   3rd Qu.:37.90  
#>  Max.   :8.780         Max.   :1.385e+09   Max.   :53.80  
#>  NA's   :42            NA's   :2           NA's   :2      
#>  Population.growth.rate   Birth.rate      Death.rate     Net.migration.rate
#>  Min.   :-3.1300        Min.   : 6.50   Min.   : 1.600   Min.   :-40.3000  
#>  1st Qu.: 0.3725        1st Qu.:12.03   1st Qu.: 6.025   1st Qu.: -2.1500  
#>  Median : 1.0050        Median :17.40   Median : 7.400   Median : -0.3000  
#>  Mean   : 1.1277        Mean   :19.66   Mean   : 7.765   Mean   : -0.6313  
#>  3rd Qu.: 1.8775        3rd Qu.:24.10   3rd Qu.: 9.400   3rd Qu.:  0.9000  
#>  Max.   : 7.3700        Max.   :43.70   Max.   :19.300   Max.   : 57.0000  
#>  NA's   :3              NA's   :3       NA's   :3        NA's   :2         
#>  Maternal.mortality.rate Infant.mortality.rate Life.expectancy.at.birth
#>  Min.   :   2.0          Min.   :  1.60        Min.   :52.10           
#>  1st Qu.:  12.5          1st Qu.:  7.20        1st Qu.:66.83           
#>  Median :  58.0          Median : 14.70        Median :74.30           
#>  Mean   : 164.9          Mean   : 23.42        Mean   :72.47           
#>  3rd Qu.: 205.5          3rd Qu.: 34.40        3rd Qu.:77.53           
#>  Max.   :1150.0          Max.   :108.50        Max.   :89.40           
#>  NA's   :18              NA's   :4             NA's   :5               
#>  Total.fertility.rate Obesity.adult.prevalence.rate
#>  Min.   :0.840        Min.   : 2.100               
#>  1st Qu.:1.740        1st Qu.: 8.875               
#>  Median :2.070        Median :20.550               
#>  Mean   :2.599        Mean   :19.514               
#>  3rd Qu.:3.030        3rd Qu.:25.375               
#>  Max.   :6.350        Max.   :61.000               
#>  NA's   :4            NA's   :9                    
#>  Children.under.the.age.of.5.years.underweight Education.expenditures
#>  Min.   : 0.50                                 Min.   : 1.000        
#>  1st Qu.: 3.40                                 1st Qu.: 3.200        
#>  Median :10.20                                 Median : 4.300        
#>  Mean   :12.52                                 Mean   : 4.525        
#>  3rd Qu.:18.25                                 3rd Qu.: 5.300        
#>  Max.   :39.90                                 Max.   :12.800        
#>  NA's   :74                                    NA's   :36            
#>  Unemployment.youth.ages.15.24    Airports          Railways       
#>  Min.   : 0.50                 Min.   :    1.0   Min.   :     9.0  
#>  1st Qu.: 8.70                 1st Qu.:   14.0   1st Qu.:   720.8  
#>  Median :14.80                 Median :   44.0   Median :  2504.0  
#>  Mean   :17.55                 Mean   :  216.8   Mean   :  9940.0  
#>  3rd Qu.:24.18                 3rd Qu.:  117.0   3rd Qu.:  5568.5  
#>  Max.   :53.40                 Max.   :13513.0   Max.   :293564.0  
#>  NA's   :31                    NA's   :6         NA's   :65        
#>     Roadways       Merchant.marine    Capital         
#>  Min.   :      8   Min.   :   1.0   Length:197        
#>  1st Qu.:   7000   1st Qu.:  28.0   Class :character  
#>  Median :  34903   Median : 100.0   Mode  :character  
#>  Mean   : 208559   Mean   : 548.9                     
#>  3rd Qu.: 104000   3rd Qu.: 417.0                     
#>  Max.   :6586610   Max.   :9053.0                     
#>  NA's   :16        NA's   :36