Montgomery County, Maryland car crash data

This dataset provides general information about each collision and details of all traffic collisions occurring on county and local roadways within Montgomery County between January 2014 and October 2018, as collected via the Automated Crash Reporting System (ACRS) of the Maryland State Police, and reported by the Montgomery County Police, Gaithersburg Police, Rockville Police, or the Maryland-National Capital Park Police.

Format

A data frame with 44379 rows and 23 variables:

Crash_Type

Identifies crash as property, injury, or fatal.

Route_Type

Type of roadway at crash location.

Cross_Street_Type

Roadway type for nearest cross-street.

Municipality

Jurisdiction for crash location.

Related_NonMotorist

Type(s) of non-motorist involved.

At_Fault

Fault by a driver or non-motorist.

Collision_Type

Type of collision.

Weather

Weather at collision location.

Surface_Condition

Condition of roadway surface.

Brightness

Lighting conditions.

Traffic_Control

Signage or traffic control devices.

Driver_Substance_Abuse

Substance abuse detected for all drivers involved.

NonMotorist_Substance_Abuse

Substance abuse detected for all non-motorists involved.

First_Collision_Event

The first event of the collision.

Second_Collision_Event

Notes is a second collision event took place.

Fixed_Object_Struck

The fixed object struck by vehicle (if applicable).

Junction

The type of junction where the collision occurred.

Intersection_Type

If the collision was intersection related, this field describes the intersection characteristics.

Road_Alignment

The road alignment where the collision occurred.

Road_Condition

The condition of the road when the collision occurred.

Road_Division

The road division where the collision occurred.

Latitude

Y coordinate of crash location.

Longitude

X coordinate of crash location.

Source

Montgomery Car Crash Data