Two-way frequency table

This function creates a two way frequency table.

crosstab(
  data,
  rowvar,
  colvar,
  type = c("freq", "cellpercent", "rowpercent", "colpercent"),
  total = TRUE,
  na.rm = TRUE,
  digits = 2,
  plot = FALSE
)

Arguments

data
data frame

rowvar
row factor (unquoted)

colvar
column factor (unquoted)

type
statistics to print. Options are "freq", "cellpercent", "rowpercent", or "colpercent" for frequencies, cell percents, row percents, or column percents). Default: "freq".

total
if TRUE, includes total percents, Default: TRUE

na.rm
if TRUE, deletes cases with missing values, Default: TRUE

digits
number of decimal digits to report for percents, Default: 2

plot
logical. If TRUE generate stacked bar chart. Default: FALSE.

Value

a list with 9 components:

tbl (table). Table of frequencies

cellPcts (table). Table of cell percents

rowPcts (table). Table of row percents

colPcts (table). Table of column percents

type (character). Type of table to print

total (logical). If TRUE, print row and or column totals

digits (numeric). number of digits to print

rowname (character). Row variable name

colname (character). Column variable name

Details

This function calculates and prints a two way frequency table. Given a data frame, a row factor, a column factor, and a type (frequencies, cell percents, row percents, or column percents) the function prints a table with labeled rows and columns, frequencies or percents with the percentage sign depending on type, a level labeled <NA> if na.rm = FALSE, and a level labeled Total if total = TRUE.

Examples

crosstab(mtcars, cyl, gear)

#>        gear
#> cyl      3  4  5 Total
#>   4      1  8  2    11
#>   6      2  4  1     7
#>   8     12  0  2    14
#>   Total 15 12  5    32

crosstab(mtcars, cyl, gear, type="cellpercent")

#>        gear
#> cyl           3       4       5   Total
#>   4       3.12%  25.00%   6.25%  34.38%
#>   6       6.25%  12.50%   3.12%  21.88%
#>   8      37.50%   0.00%   6.25%  43.75%
#>   Total  46.88%  37.50%  15.62% 100.00%

crosstab(mtcars, cyl, gear, type="colpercent")

#>        gear
#> cyl           3       4       5
#>   4       6.67%  66.67%  40.00%
#>   6      13.33%  33.33%  20.00%
#>   8      80.00%   0.00%  40.00%
#>   Total 100.00% 100.00% 100.00%

crosstab(mtcars, cyl, gear, type="rowpercent")

#>    gear
#> cyl       3       4       5   Total
#>   4   9.09%  72.73%  18.18% 100.00%
#>   6  28.57%  57.14%  14.29% 100.00%
#>   8  85.71%   0.00%  14.29% 100.00%