df_summary

A comprehensive description of a data frame

df_summary(df, digits = 2, maxcat = 10, label_length = 40)

Arguments

df
a data frame

digits
number of decimal digits for statistics.

maxcat
maximum number of levels of a character/factor variable to print.

label_length
maximum length of factor level label to print. Longer labels will be truncated.

Value

the data frame invisibly (so that it can be used in a pipeline)

Details

Prints a comprehensive description of a data frame via several tables, a general summary table and tables that provide a breakdown of quantitative and categorical variables

Examples

testdata <- data.frame(height=c(4, 5, 3, 2, 100),
                       weight=c(39, 88, NA, 15, -2),
                       names=c("Bill","Dean", "Sam", NA, "Jane"),
                       race=c('b', 'w', 'w', 'o', 'b'))

df_summary(testdata)

#> 
#> The data frame testdata has 5 observations and 4 variables.
#> 
#>  Overall 
#>  pos variable      type n_unique n_missing pct_missing
#>    1   height   numeric        5         0          0%
#>    2   weight   numeric        5         1         20%
#>    3    names character        5         1         20%
#>    4     race character        3         0          0%
#> 
#>  Quantitative Variables 
#>    name n mean    sd skew min   p25 median   p75 max
#>  height 5 22.8 43.17 1.07   2  3.00      4  5.00 100
#>  weight 4 35.0 39.13 0.38  -2 10.75     27 51.25  88
#> 
#>  Categorical Variables 
#>  variable level n pct 
#>  names    Bill  1 0.25
#>           Dean  1 0.25
#>           Jane  1 0.25
#>           Sam   1 0.25
#>  race     b     2 0.40
#>           o     1 0.20
#>           w     2 0.40