Google Play Store Apps

A data frame containing all the details of the applications on Google Play. There are 13 features that describe a given app.

googleplay

Format

A data frame with 10,842 rows and 13 variables The variables are as follows:

app

application name

category

category the app belongs to ss

rating

overall user rating of the app

reviews

number of user reviews for the app

size

size of the app

installs

number of user downloads/installs for the app

type

paid or free

price

price of the app

content_rating

age group the app is targeted at

genres

an app can belong to multiple genres apart from its main category

last_updated

date when the app was last updated on play store

current_ver

current version of the app available on play store

android_ver

min required android version

Source

The data is from Kaggle. Lavanya Gupta created the Google Play Store Apps on Kaggle. https://www.kaggle.com/lava18/google-play-store-apps

Details

Lavanya Gupta created a public dataset on Kaggle about Google Play Store applications. The information in the dataset was scraped from the Google Play Store. Because there are apps always being added or removed from Google Play Store, Gupta has modified this dataset over time. This dataset is the most recent version, Version 5, and was uploaded by Gupta to Kaggle on September 2018.

Note

The dataset has some challenges for students. For example, the size variable needs to be converted to a standard numeric unit. The genres field can contain more than one value seperated by semicolons. The number of installs needs to be converted to an ordinal factor. The type factor has an NaN value.

Examples

summary(googleplay)

#>      app                    category        rating         reviews        
#>  Length:10840       FAMILY      :1972   Min.   :1.000   Min.   :       0  
#>  Class :character   GAME        :1144   1st Qu.:4.000   1st Qu.:      38  
#>  Mode  :character   TOOLS       : 843   Median :4.300   Median :    2094  
#>                     MEDICAL     : 463   Mean   :4.192   Mean   :  444153  
#>                     BUSINESS    : 460   3rd Qu.:4.500   3rd Qu.:   54776  
#>                     PRODUCTIVITY: 424   Max.   :5.000   Max.   :78158306  
#>                     (Other)     :5534   NA's   :1474                      
#>      size             installs           type           price        
#>  Length:10840       Length:10840       Free:10039   Min.   :  0.000  
#>  Class :character   Class :character   NaN :    1   1st Qu.:  0.000  
#>  Mode  :character   Mode  :character   Paid:  800   Median :  0.000  
#>                                                     Mean   :  1.027  
#>                                                     3rd Qu.:  0.000  
#>                                                     Max.   :400.000  
#>                                                                      
#>          content_rating    genres           last_updated                
#>  Adults only 18+:   3   Length:10840       Min.   :2010-05-21 00:00:00  
#>  Everyone       :8714   Class :character   1st Qu.:2017-09-20 00:00:00  
#>  Everyone 10+   : 414   Mode  :character   Median :2018-05-24 00:00:00  
#>  Mature 17+     : 499                      Mean   :2017-11-21 06:43:02  
#>  Teen           :1208                      3rd Qu.:2018-07-20 00:00:00  
#>  Unrated        :   2                      Max.   :2018-08-08 00:00:00  
#>                                                                         
#>  current_ver                    android_ver  
#>  Length:10840       4.1 and up        :2451  
#>  Class :character   4.0.3 and up      :1501  
#>  Mode  :character   4.0 and up        :1375  
#>                     Varies with device:1362  
#>                     4.4 and up        : 980  
#>                     2.3 and up        : 652  
#>                     (Other)           :2519  

table(googleplay$category,googleplay$type)

#>                      
#>                       Free  NaN Paid
#>   ART_AND_DESIGN        62    0    3
#>   AUTO_AND_VEHICLES     82    0    3
#>   BEAUTY                53    0    0
#>   BOOKS_AND_REFERENCE  203    0   28
#>   BUSINESS             446    0   14
#>   COMICS                60    0    0
#>   COMMUNICATION        360    0   27
#>   DATING               227    0    7
#>   EDUCATION            152    0    4
#>   ENTERTAINMENT        147    0    2
#>   EVENTS                63    0    1
#>   FAMILY              1780    1  191
#>   FINANCE              349    0   17
#>   FOOD_AND_DRINK       125    0    2
#>   GAME                1061    0   83
#>   HEALTH_AND_FITNESS   325    0   16
#>   HOUSE_AND_HOME        88    0    0
#>   LIBRARIES_AND_DEMO    84    0    1
#>   LIFESTYLE            363    0   19
#>   MAPS_AND_NAVIGATION  132    0    5
#>   MEDICAL              354    0  109
#>   NEWS_AND_MAGAZINES   281    0    2
#>   PARENTING             58    0    2
#>   PERSONALIZATION      309    0   83
#>   PHOTOGRAPHY          313    0   22
#>   PRODUCTIVITY         396    0   28
#>   SHOPPING             258    0    2
#>   SOCIAL               292    0    3
#>   SPORTS               360    0   24
#>   TOOLS                765    0   78
#>   TRAVEL_AND_LOCAL     246    0   12
#>   VIDEO_PLAYERS        171    0    4
#>   WEATHER               74    0    8