Holland Occupational Themes

RIASEC/Holland Code test

holland

Format

A data frame with 18231 rows and 68 variables:

R1

Test the quality of parts before shipment

R2

Lay brick or tile

R3

Work on an offshore oil-drilling rig

R4

Assemble electronic parts

R5

Operate a grinding machine in a factory

R6

Fix a broken faucet

R7

Assemble products in a factory

R8

Install flooring in houses

I1

Study the structure of the human body

I2

Study animal behavior

I3

Do research on plants or animals

I4

Develop a new medical treatment or procedure

I5

Conduct biological research

I6

Study whales and other types of marine life

I7

Work in a biology lab

I8

Make a map of the bottom of an ocean

A1

Conduct a musical choir

A2

Direct a play

A3

Design artwork for magazines

A4

Write a song

A5

Write books or plays

A6

Play a musical instrument

A7

Perform stunts for a movie or television show

A8

Design sets for plays

S1

Give career guidance to people

S2

Do volunteer work at a non-profit organization

S3

Help people who have problems with drugs or alcohol

S4

Teach an individual an exercise routine

S5

Help people with family-related problems

S6

Supervise the activities of children at a camp

S7

Teach children how to read

S8

Help elderly people with their daily activities

E1

Sell restaurant franchises to individuals

E2

Sell merchandise at a department store

E3

Manage the operations of a hotel

E4

Operate a beauty salon or barber shop

E5

Manage a department within a large company

E6

Manage a clothing store

E7

Sell houses

E8

Run a toy store

C1

Generate the monthly payroll checks for an office

C2

Inventory supplies using a hand-held computer

C3

Use a computer program to generate customer bills

C4

Maintain employee records

C5

Compute and record statistical and other numerical data

C6

Operate a calculator

C7

Handle customers' bank transactions

C8

Keep shipping and receiving records

TIPI1

Extraverted, enthusiastic.

TIPI2

Critical, quarrelsome.

TIPI3

Dependable, self-disciplined.

TIPI4

Anxious, easily upset.

TIPI5

Open to new experiences, complex.

TIPI6

Reserved, quiet.

TIPI7

Sympathetic, warm.

TIPI8

Disorganized, careless.

TIPI9

Calm, emotionally stable.

TIPI10

Conventional, uncreative.

education

factor. How much education have you completed?

urban

factor. What type of area did you live when you were a child?

gender

factor. What is your gender?

age

integer. Age in years.

hand

factor. What hand do you use to write with?

religion

factor. What is your religion?

voted

factor. Have you voted in a national election in the past year?

married

factor. What is your marital status?.

familysize

integer. Including you, how many children did your mother have?

major

factor. If you attended a university, what was your major?

Source

Dataset obtained from the Open-Source Psychometrics Project.

Details

This is a nonrandom subset of the larger online data collected between 2015-2018. The sample has been restricted to a US sample with the most commonly identified majors and religions. Not all variables were selected.

The RIASEC has 48 items, rated on a 1-5 scale of how much they would like to perform that task, with the labels 1=Dislike, 3=Neutral, 5=Enjoy. The form 6 scales:

Realistic

Doer - prefers technical, outdoor and athletic pursuits

Investigative

Thinker - prefers scientific, research and intellectual pursuits

Artistic

Creator - prefers creative, imaginative and intuitive pursuits

Social

Helper - prefers helping, and developing and interpersonal pursuits

Enterprising

Persuader - prefers leadership, influencing and persuasive pursuits

Conventional

Organizer - prefers data management, numerical and organizational pursuits

TIPI1 through TIPI10 are the Ten Item Personality Inventory (see Gosling, S. D., Rentfrow, P. J., & Swann, W. B., Jr. (2003). A Very Brief Measure of the Big Five Personality Domains. Journal of Research in Personality, 37, 504-528.).

The TIPI items were rated "I see myself as:" _____ from 1 = Disagree strongly to 7=Agree strongly.

Examples

summary(holland)

#>        R1              R2              R3              R4       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000  
#>  Median :2.000   Median :2.000   Median :1.000   Median :2.000  
#>  Mean   :2.335   Mean   :2.001   Mean   :1.548   Mean   :2.045  
#>  3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:2.000   3rd Qu.:3.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        R5              R6              R7              R8              I1      
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.00  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:3.00  
#>  Median :1.000   Median :2.000   Median :1.000   Median :1.000   Median :4.00  
#>  Mean   :1.569   Mean   :2.119   Mean   :1.801   Mean   :1.889   Mean   :3.59  
#>  3rd Qu.:2.000   3rd Qu.:3.000   3rd Qu.:2.000   3rd Qu.:3.000   3rd Qu.:5.00  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.00  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1       NA's   :1     
#>        I2              I3              I4              I5       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:3.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :4.000   Median :3.000   Median :3.000   Median :3.000  
#>  Mean   :3.418   Mean   :3.091   Mean   :3.162   Mean   :2.917  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        I6              I7              I8             A1              A2       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.00   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:2.000   1st Qu.:1.000   1st Qu.:1.00   1st Qu.:1.000   1st Qu.:1.000  
#>  Median :3.000   Median :3.000   Median :2.00   Median :2.000   Median :3.000  
#>  Mean   :2.999   Mean   :2.784   Mean   :2.44   Mean   :2.352   Mean   :2.636  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.00   3rd Qu.:3.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.00   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1      NA's   :1       NA's   :1      
#>        A3              A4              A5             A6              A7      
#>  Min.   :0.000   Min.   :0.000   Min.   :0.00   Min.   :0.000   Min.   :0.00  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:2.00   1st Qu.:2.000   1st Qu.:1.00  
#>  Median :3.000   Median :3.000   Median :3.00   Median :3.000   Median :2.00  
#>  Mean   :2.859   Mean   :2.894   Mean   :3.03   Mean   :3.099   Mean   :2.54  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.00   3rd Qu.:4.000   3rd Qu.:4.00  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.00   Max.   :5.000   Max.   :5.00  
#>  NA's   :1       NA's   :1       NA's   :1      NA's   :1       NA's   :1     
#>        A8              S1              S2              S3       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:3.000   1st Qu.:3.000   1st Qu.:3.000  
#>  Median :3.000   Median :4.000   Median :4.000   Median :4.000  
#>  Mean   :2.679   Mean   :3.753   Mean   :3.847   Mean   :3.588  
#>  3rd Qu.:4.000   3rd Qu.:5.000   3rd Qu.:5.000   3rd Qu.:5.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        S4              S5              S6              S7       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:2.000   1st Qu.:3.000   1st Qu.:2.000   1st Qu.:3.000  
#>  Median :3.000   Median :4.000   Median :3.000   Median :4.000  
#>  Mean   :3.233   Mean   :3.907   Mean   :3.218   Mean   :3.443  
#>  3rd Qu.:4.000   3rd Qu.:5.000   3rd Qu.:4.000   3rd Qu.:5.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        S8              E1              E2              E3       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:2.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000  
#>  Median :3.000   Median :2.000   Median :2.000   Median :3.000  
#>  Mean   :3.021   Mean   :2.025   Mean   :2.291   Mean   :2.679  
#>  3rd Qu.:4.000   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        E4              E5              E6              E7              E8      
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.00  
#>  1st Qu.:1.000   1st Qu.:2.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.00  
#>  Median :2.000   Median :3.000   Median :2.000   Median :2.000   Median :2.00  
#>  Mean   :2.344   Mean   :2.935   Mean   :2.474   Mean   :2.533   Mean   :2.53  
#>  3rd Qu.:3.000   3rd Qu.:4.000   3rd Qu.:3.000   3rd Qu.:4.000   3rd Qu.:4.00  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.00  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1       NA's   :1     
#>        C1             C2              C3              C4              C5       
#>  Min.   :0.00   Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:1.00   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000  
#>  Median :2.00   Median :2.000   Median :2.000   Median :2.000   Median :2.000  
#>  Mean   :2.31   Mean   :2.297   Mean   :2.297   Mean   :2.497   Mean   :2.525  
#>  3rd Qu.:3.00   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:4.000  
#>  Max.   :5.00   Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1      NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        C6              C7              C8            TIPI1      
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:3.000  
#>  Median :3.000   Median :2.000   Median :2.000   Median :5.000  
#>  Mean   :2.643   Mean   :2.207   Mean   :2.154   Mean   :4.716  
#>  3rd Qu.:4.000   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:6.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :7.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>      TIPI2           TIPI3           TIPI4           TIPI5      
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:2.000   1st Qu.:5.000   1st Qu.:2.000   1st Qu.:5.000  
#>  Median :4.000   Median :6.000   Median :4.000   Median :6.000  
#>  Mean   :3.663   Mean   :5.859   Mean   :3.904   Mean   :5.766  
#>  3rd Qu.:5.000   3rd Qu.:7.000   3rd Qu.:5.000   3rd Qu.:7.000  
#>  Max.   :7.000   Max.   :7.000   Max.   :7.000   Max.   :7.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>      TIPI6           TIPI7           TIPI8         TIPI9           TIPI10     
#>  Min.   :0.000   Min.   :0.000   Min.   :0.0   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:3.000   1st Qu.:5.000   1st Qu.:1.0   1st Qu.:4.000   1st Qu.:1.000  
#>  Median :5.000   Median :6.000   Median :2.0   Median :6.000   Median :2.000  
#>  Mean   :4.453   Mean   :5.843   Mean   :2.7   Mean   :5.232   Mean   :2.831  
#>  3rd Qu.:6.000   3rd Qu.:7.000   3rd Qu.:4.0   3rd Qu.:6.000   3rd Qu.:4.000  
#>  Max.   :7.000   Max.   :7.000   Max.   :7.0   Max.   :7.000   Max.   :7.000  
#>  NA's   :1       NA's   :1       NA's   :1     NA's   :1       NA's   :1      
#>                  education                     urban         gender     
#>  Less than high school: 770   Rural (country side):4226   Male  : 5029  
#>  High school          :6467   Suburban            :7540   Female:13201  
#>  University degree    :7215   Urban (town, city)  :6362   NA's  :    1  
#>  Graduate degree      :3679   NA's                : 103                 
#>  NA's                 : 100                                             
#>                                                                         
#>                                                                         
#>       age           hand                         religion     voted     
#>  Min.   :13.00   Right:16013   Christian (Other)     :5069   Yes :9179  
#>  1st Qu.:20.00   Left : 1834   Christian (Catholic)  :4084   No  :8883  
#>  Median :25.00   Both :  316   Christian (Protestant):3063   NA's: 169  
#>  Mean   :29.53   NA's :   68   Agnostic              :2576              
#>  3rd Qu.:37.00                 Atheist               :1660              
#>  Max.   :87.00                 (Other)               :1504              
#>  NA's   :6                     NA's                  : 275              
#>                married        familysize            major     
#>  Never married     :11665   Min.   : 0.000   psychology:7745  
#>  Currently married : 5107   1st Qu.: 2.000   business  :2103  
#>  Previously married: 1375   Median : 3.000   nursing   :1771  
#>  NA's              :   84   Mean   : 2.873   biology   :1315  
#>                             3rd Qu.: 3.000   english   :1299  
#>                             Max.   :27.000   (Other)   :3997  
#>                             NA's   :4        NA's      :   1