UK oil investments between 1969 and 1992

Dataset on oil investment

oil

Format

A data frame with 53 rows and 12 variables:

dur

duration of the appraisal lag in months (time span between discovery of an oil field and beginning of development, i.e. approval of annex

size

size of recoverable reserves in millions of barrels

waterd

depth of the sea in metres

gasres

size of recoverable gas reserves in billions of cubic feet

operator

equity market value (in 1991 million pounds) of the company operating the oil field

p

real after–tax oil price measured at time of annex B approval

vardp

volatility of the real oil price process measured as the squared recursive standard errors of the regression of pt-pt-1 on a constant

p97

adaptive expectations (with parameter theta=0.97) for the real after–tax oil prices formed at the time of annex B approval

varp97

volatility of the adaptive expectations (with parameter theta=0.97) for real after tax oil prices measured as the squared recursive standard errors of the regression of pt on pte(theta)

p98

adaptive expectations (with parameter theta=0.98) for the real after–tax oil prices formed at the time of annex B approval

varp98

volatility of the adaptive expectations (with parameter theta=0.98) for real after tax oil prices measured as the squared recursive standard errors of the regression of pt on pte(theta)

Source

Datasets from vincentarelbundock on github.

Details

Data for oil investments between the period 1969 to 1992 in the United Kingdom.

Examples

summary(oil)

#>       dur              size            waterd          gasres      
#>  Min.   :  1.00   Min.   :   4.0   Min.   : 76.0   Min.   :   0.0  
#>  1st Qu.: 16.00   1st Qu.:  60.0   1st Qu.:108.0   1st Qu.:   0.0  
#>  Median : 51.00   Median : 110.0   Median :135.0   Median :   0.0  
#>  Mean   : 63.02   Mean   : 328.9   Mean   :126.5   Mean   : 178.8  
#>  3rd Qu.:102.00   3rd Qu.: 300.0   3rd Qu.:147.0   3rd Qu.:  50.0  
#>  Max.   :227.00   Max.   :3030.0   Max.   :186.0   Max.   :2500.0  
#>     operator           p             vardp            p97        
#>  Min.   :    1   Min.   :0.921   Min.   :0.007   Min.   :0.8979  
#>  1st Qu.: 3173   1st Qu.:1.284   1st Qu.:1.113   1st Qu.:1.2551  
#>  Median :16000   Median :2.243   Median :1.852   Median :1.8843  
#>  Mean   :11522   Mean   :2.030   Mean   :1.544   Mean   :1.6962  
#>  3rd Qu.:16175   3rd Qu.:2.639   3rd Qu.:2.157   3rd Qu.:2.0480  
#>  Max.   :17928   Max.   :3.366   Max.   :2.425   Max.   :2.1581  
#>      varp97           p98             varp98     
#>  Min.   :0.150   Min.   :0.8918   Min.   :0.155  
#>  1st Qu.:2.401   1st Qu.:1.3565   1st Qu.:2.474  
#>  Median :3.255   Median :1.9763   Median :3.736  
#>  Mean   :2.979   Mean   :1.7860   Mean   :3.232  
#>  3rd Qu.:3.494   3rd Qu.:2.2089   3rd Qu.:3.905  
#>  Max.   :5.037   Max.   :2.2835   Max.   :5.238