Soil composition data

Soil Compositions of Physical and Chemical Characteristics

soils

Format

A data frame with 48 rows and 14 variables:

Group

grouping variable with 12 levels, corresponding to the combinations of Contour and Depth

Contour

character variable with 3 levels: Depression Slope Top

Depth

character variable with 4 levels: 0-10 10-30 30-60 60-90

Gp

character variable with 12 levels, giving abbreviations for the groups: D0 D1 D3 D6 S0 S1 S3 S6 T0 T1 T3 T6

Block

a factor with levels 1 2 3 4

pH

soil pH

N

total nitrogen in %

Dens

bulk density in gm/cm$^3$

P

total phosphorous in ppm

Ca

calcium in me/100 gm.

Mg

magnesium in me/100 gm.

K

phosphorous in me/100 gm.

Na

sodium in me/100 gm.

Conduc

conductivity

Source

Rdatasets from vincentarelbundock github

Details

Soil characteristics as measured on samples from three types of contours (Top, Slope, and Depression) and at four depths (0-10cm, 10-30cm, 30-60cm, and 60-90cm). The area was divided into 4 blocks, in a randomized block design. (Suggested by Michael Friendly.)