Major sports venues

The data is a subset of the data extracted from HIFLD Open Data. It includes all of the major venues for American professional sports, their location, and what type of sports are performed there.

venues

Format

A data frame with 824 rows and 8 variables:

venueid

Unique ID for venue

name

Venue Name

city

Venue's city

state

Venue's state

zip

Venue's Zip code

status

State of venue: CLOSED, OPEN or UNDER CONSTRUCTION

country

Country of Venue

type

NAICS Code description of venue

Source

Homeland Infrastructure Foundation-Level Data

Examples

summary(venues)

#>     venueid           name               city               state    
#>  Min.   :   1.0   Length:824         Length:824         CA     : 70  
#>  1st Qu.: 251.8   Class :character   Class :character   TX     : 59  
#>  Median : 912.5   Mode  :character   Mode  :character   FL     : 43  
#>  Mean   : 741.0                                         NY     : 43  
#>  3rd Qu.:1125.2                                         NC     : 34  
#>  Max.   :1336.0                                         OH     : 34  
#>                                                         (Other):541  
#>       zip                     status    country  
#>  02215  :  3   CLOSED            :  6   can: 17  
#>  19148  :  3   OPEN              :813   mex: 21  
#>  28202  :  3   UNDER CONSTRUCTION:  5   pri:  1  
#>  37212  :  3                            usa:785  
#>  38677  :  3                                     
#>  44115  :  3                                     
#>  (Other):806                                     
#>                                                                        type    
#>  GOLF COURSES AND COUNTRY CLUBS                                          : 70  
#>  PROMOTERS OF PERFORMING ARTS, SPORTS, AND SIMILAR EVENTS WITH FACILITIES:680  
#>  RACETRACKS                                                              : 74  
#>                                                                                
#>                                                                                
#>                                                                                
#>