Forest fires in Brazil

This dataset reports the number of forest fires in Brazil per state. The series has data from a 10 year period from 1998 to 2017.

Format

A data frame with 6454 rows and 5 variables:

year

Year when forest fire(s) occured

state

Name of the state of Brazil where the forest fire(s) occured

month

Month when the forest fire(s)

number

Number of forest fires reported

date

Date when forest fire(s) records were made

Source

Sistema Nacional de Informacoes Florestais