Soil composition data

Soil Compositions of Physical and Chemical Characteristics

soils

Format

A data frame with 48 rows and 14 variables:

Group

grouping variable with 12 levels, corresponding to the combinations of Contour and Depth

Contour

character variable with 3 levels: Depression Slope Top

Depth

character variable with 4 levels: 0-10 10-30 30-60 60-90

Gp

character variable with 12 levels, giving abbreviations for the groups: D0 D1 D3 D6 S0 S1 S3 S6 T0 T1 T3 T6

Block

a factor with levels 1 2 3 4

pH

soil pH

N

total nitrogen in %

Dens

bulk density in gm/cm$^3$

P

total phosphorous in ppm

Ca

calcium in me/100 gm.

Mg

magnesium in me/100 gm.

K

phosphorous in me/100 gm.

Na

sodium in me/100 gm.

Conduc

conductivity

Source

Rdatasets from vincentarelbundock github

Details

Soil characteristics as measured on samples from three types of contours (Top, Slope, and Depression) and at four depths (0-10cm, 10-30cm, 30-60cm, and 60-90cm). The area was divided into 4 blocks, in a randomized block design. (Suggested by Michael Friendly.)

Examples

summary(soils)

#>      Group      Contour             Depth                Gp            Block 
#>  1      : 4   Length:48          Length:48          Length:48          1:12  
#>  2      : 4   Class :character   Class :character   Class :character   2:12  
#>  3      : 4   Mode  :character   Mode  :character   Mode  :character   3:12  
#>  4      : 4                                                            4:12  
#>  5      : 4                                                                  
#>  6      : 4                                                                  
#>  (Other):24                                                                  
#>        pH              N                Dens             P        
#>  Min.   :3.740   Min.   :0.03000   Min.   :0.780   Min.   : 79.0  
#>  1st Qu.:4.058   1st Qu.:0.05075   1st Qu.:1.127   1st Qu.:108.8  
#>  Median :4.545   Median :0.08450   Median :1.400   Median :131.0  
#>  Mean   :4.669   Mean   :0.10194   Mean   :1.316   Mean   :166.2  
#>  3rd Qu.:5.140   3rd Qu.:0.12925   3rd Qu.:1.502   3rd Qu.:214.2  
#>  Max.   :6.670   Max.   :0.29800   Max.   :1.600   Max.   :445.0  
#>                                                                   
#>        Ca               Mg               K                Na        
#>  Min.   : 3.820   Min.   : 5.150   Min.   :0.1400   Min.   : 0.600  
#>  1st Qu.: 5.040   1st Qu.: 7.537   1st Qu.:0.2750   1st Qu.: 2.545  
#>  Median : 7.305   Median : 8.515   Median :0.4250   Median : 5.520  
#>  Mean   : 8.029   Mean   : 8.465   Mean   :0.4662   Mean   : 5.600  
#>  3rd Qu.: 9.735   3rd Qu.: 9.648   3rd Qu.:0.6425   3rd Qu.: 8.355  
#>  Max.   :16.350   Max.   :10.960   Max.   :1.0900   Max.   :11.040  
#>                                                                     
#>      Conduc      
#>  Min.   : 0.670  
#>  1st Qu.: 2.790  
#>  Median : 6.635  
#>  Mean   : 6.589  
#>  3rd Qu.: 9.852  
#>  Max.   :13.320  
#>